Changed arround some files to make pong a bit more readable, need to implement physics again
