--- .vscode/launch.json
@@ -1,30 +1,0 @@
-{
-	"configurations": [
-		{
-			"address": "localhost",
-			"localRoot": "${workspaceFolder}/MVC",
-			"name": "Attach to MVC",
-			"port": 9321,
-			"remoteRoot": "/app",
-			"request": "attach",
-			"skipFiles": [
-				"<node_internals>/**",
-				"${workspaceFolder}/MVC/node_modules/**"
-			],
-			"type": "node"
-		},
-		{
-			"address": "localhost",
-			"localRoot": "${workspaceFolder}/API",
-			"name": "Attach to API",
-			"port": 9320,
-			"remoteRoot": "/app",
-			"request": "attach",
-			"skipFiles": [
-				"<node_internals>/**",
-				"${workspaceFolder}/API/node_modules/**"
-			],
-			"type": "node"
-		}
-	]
-}
